+++ src/lib/server/db/actions/importActions.ts
@@ -12,28 +12,31 @@ import {
 	importTable,
 	journalEntry,
 	label,
-	tag,
-	transaction
+	tag
 } from '../schema';
 import { updatedTime } from './helpers/updatedTime';
 import { eq, and, getTableColumns, desc, sql } from 'drizzle-orm';
 import Papa from 'papaparse';
-import {
-	createCombinedTransactionSchema,
-	createSimpleTransactionSchema,
-	type CreateCombinedTransactionType,
-	type CreateSimpleTransactionType
-} from '$lib/schema/journalSchema';
+import { createSimpleTransactionSchema } from '$lib/schema/journalSchema';
 import { tActions } from './tActions';
 import { filterNullUndefinedAndDuplicates } from '../../../../routes/(loggedIn)/journals/filterNullUndefinedAndDuplicates';
 import { createAccountSchema } from '$lib/schema/accountSchema';
-import type { ZodSchema, ZodType } from 'zod';
+import type { ZodSchema } from 'zod';
 import { createBillSchema } from '$lib/schema/billSchema';
 import { createBudgetSchema } from '$lib/schema/budgetSchema';
 import { createCategorySchema } from '$lib/schema/categorySchema';
 import { createTagSchema } from '$lib/schema/tagSchema';
 import { createLabelSchema } from '$lib/schema/labelSchema';
 import { importTypeEnum, type importTypeType } from '$lib/schema/importSchema';
+import {
+	importTransaction,
+	importAccount,
+	importBill,
+	importBudget,
+	importCategory,
+	importTag,
+	importLabel
+} from './helpers/importHelpers';
 
 export const importActions = {
 	list: async ({ db }: { db: DBType }) => {
@@ -562,328 +565,3 @@ export const importActions = {
 		}
 	}
 };
-
-const simpleSchemaToCombinedSchema = (
-	data: CreateSimpleTransactionType
-): CreateCombinedTransactionType => {
-	const {
-		toAccountId,
-		toAccountTitle,
-		fromAccountId,
-		fromAccountTitle,
-		amount,
-		...sharedProperties
-	} = data;
-
-	return [
-		{
-			...sharedProperties,
-			accountId: fromAccountId,
-			accountTitle: fromAccountTitle,
-			amount: -amount
-		},
-		{ ...sharedProperties, accountId: toAccountId, accountTitle: toAccountTitle, amount: amount }
-	];
-};
-
-const importItem = async <T extends Record<string, unknown>, DBT extends { id: string }>({
-	db,
-	item,
-	schema,
-	createItem
-}: {
-	db: DBType;
-	item: typeof importItemDetail.$inferSelect;
-	schema: ZodType<T>;
-	createItem: (data: { item: T; db: DBType }) => Promise<DBT | undefined>;
-}) => {
-	const processedItem = schema.safeParse(item.processedInfo);
-	if (processedItem.success) {
-		try {
-			const createdItem = await createItem({ item: processedItem.data, db });
-
-			if (createdItem) {
-				await db
-					.update(importItemDetail)
-					.set({
-						status: 'imported',
-						importInfo: createdItem,
-						relationId: createdItem.id,
-						...updatedTime()
-					})
-					.where(eq(importItemDetail.id, item.id))
-					.execute();
-			} else {
-				await db
-					.update(importItemDetail)
-					.set({
-						status: 'importError',
-						errorInfo: 'Account Not Found',
-						...updatedTime()
-					})
-					.where(eq(importItemDetail.id, item.id))
-					.execute();
-			}
-		} catch (e) {
-			console.log(e);
-
-			await db
-				.update(importItemDetail)
-				.set({ status: 'importError', errorInfo: e, ...updatedTime() })
-				.where(eq(importItemDetail.id, item.id))
-				.execute();
-		}
-	} else {
-		await db
-			.update(importItemDetail)
-			.set({ status: 'importError', errorInfo: processedItem.error, ...updatedTime() })
-			.where(eq(importItemDetail.id, item.id))
-			.execute();
-	}
-};
-
-async function importTransaction({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) {
-	const processedItem = createSimpleTransactionSchema.safeParse(item.processedInfo);
-	if (processedItem.success) {
-		const combinedTransaction = simpleSchemaToCombinedSchema({
-			...processedItem.data,
-			importId: item.importId,
-			importDetailId: item.id
-		});
-		const processedCombinedTransaction =
-			createCombinedTransactionSchema.safeParse(combinedTransaction);
-		if (processedCombinedTransaction.success) {
-			try {
-				const importedData = await tActions.journal.createManyTransactionJournals({
-					db: trx,
-					journalEntries: [processedCombinedTransaction.data]
-				});
-
-				await Promise.all(
-					importedData.map(async (transactionId) => {
-						const journalData = await trx.query.transaction.findFirst({
-							where: eq(transaction.id, transactionId),
-							with: { journals: true }
-						});
-
-						if (journalData) {
-							await trx
-								.update(importItemDetail)
-								.set({
-									status: 'imported',
-									importInfo: journalData,
-									relationId: journalData.journals[0].id,
-									relation2Id: journalData.journals[1].id,
-									...updatedTime()
-								})
-								.where(eq(importItemDetail.id, item.id))
-								.execute();
-						} else {
-							await trx
-								.update(importItemDetail)
-								.set({
-									status: 'importError',
-									errorInfo: 'Journal Not Found',
-									...updatedTime()
-								})
-								.where(eq(importItemDetail.id, item.id))
-								.execute();
-						}
-					})
-				);
-			} catch (e) {
-				console.log(e);
-
-				await trx
-					.update(importItemDetail)
-					.set({ status: 'importError', errorInfo: e, ...updatedTime() })
-					.where(eq(importItemDetail.id, item.id))
-					.execute();
-			}
-		} else {
-			await trx
-				.update(importItemDetail)
-				.set({
-					status: 'importError',
-					errorInfo: processedCombinedTransaction.error,
-					...updatedTime()
-				})
-				.where(eq(importItemDetail.id, item.id))
-				.execute();
-		}
-	} else {
-		await trx
-			.update(importItemDetail)
-			.set({ status: 'importError', errorInfo: processedItem.error, ...updatedTime() })
-			.where(eq(importItemDetail.id, item.id))
-			.execute();
-	}
-}
-
-const importAccount = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createAccountSchema,
-		createItem: async (data) => {
-			const importedData = await tActions.account.create(data.db, {
-				...data.item,
-				type: data.item.type || 'expense',
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.account.findFirst({
-				where: eq(account.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
-
-const importBill = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createBillSchema,
-		createItem: async (data) => {
-			const importedData = await tActions.bill.create(data.db, {
-				...data.item,
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.bill.findFirst({
-				where: eq(bill.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
-
-const importBudget = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createBudgetSchema,
-		createItem: async (data) => {
-			const importedData = await tActions.budget.create(data.db, {
-				...data.item,
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.budget.findFirst({
-				where: eq(budget.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
-
-const importCategory = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createCategorySchema,
-		createItem: async (data) => {
-			const importedData = await tActions.category.create(data.db, {
-				...data.item,
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.category.findFirst({
-				where: eq(category.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
-
-const importTag = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createTagSchema,
-		createItem: async (data) => {
-			const importedData = await tActions.tag.create(data.db, {
-				...data.item,
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.tag.findFirst({
-				where: eq(tag.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
-const importLabel = async ({
-	item,
-	trx
-}: {
-	item: typeof importItemDetail.$inferSelect;
-	trx: DBType;
-}) =>
-	importItem({
-		db: trx,
-		item,
-		schema: createLabelSchema,
-		createItem: async (data) => {
-			const importedData = await tActions.label.create(data.db, {
-				...data.item,
-				status: data.item.status || 'active',
-				importId: item.importId,
-				importDetailId: item.id
-			});
-
-			const createdItem = await trx.query.label.findFirst({
-				where: eq(label.id, importedData)
-			});
-
-			return createdItem;
-		}
-	});
